[IMPROVEMENT] Users Presence on Demand
