+++ app/sagas/login.js
@@ -71,10 +71,6 @@ const registerPushToken = function* registerPushToken() {
 	yield RocketChat.registerPushToken();
 };
 
-const fetchUserPresence = function* fetchUserPresence() {
-	yield RocketChat.getUserPresence();
-};
-
 const handleLoginSuccess = function* handleLoginSuccess({ user }) {
 	try {
 		const adding = yield select(state => state.server.adding);
@@ -87,7 +83,6 @@ const handleLoginSuccess = function* handleLoginSuccess({ user }) {
 		yield fork(fetchRoles);
 		yield fork(fetchSlashCommands);
 		yield fork(registerPushToken);
-		yield fork(fetchUserPresence);
 
 		I18n.locale = user.language;
 		moment.locale(toMomentLocale(user.language));
